+++ postgres-single/postgres-single.d/local/share/cook/bin/postgres-post-init.sh
@@ -1,12 +1,20 @@
 #!/bin/sh
 
-# make sure we're not in /root else error:
+# make sure we're not in /root else this error occurs:
+#
 #  could not change directory to "/root": Permission denied
 #
 cd /tmp || exit 1
 
-# create postgres_exporter user
-sudo -u postgres /usr/local/bin/psql -c "CREATE USER postgres_exporter;" || true
-sudo -u postgres /usr/local/bin/psql -c "ALTER USER postgres_exporter SET SEARCH_PATH TO postgres_exporter,pg_catalog;" || true
-sudo -u postgres /usr/local/bin/psql -c "GRANT CONNECT ON DATABASE postgres TO postgres_exporter;" || true
-sudo -u postgres /usr/local/bin/psql -c "GRANT pg_monitor to postgres_exporter;" || true
+# Check if postgres_exporter user exists
+USER_EXISTS=$(sudo -u postgres /usr/local/bin/psql -tAc "SELECT 1 FROM pg_roles WHERE rolname='postgres_exporter'")
+
+# If the postgres_exporter user doesn't exist, then create and configure it
+if [ -z "$USER_EXISTS" ]; then
+    sudo -u postgres /usr/local/bin/psql -c "CREATE USER postgres_exporter with encrypted password '$EXPORTERPASS';"
+    sudo -u postgres /usr/local/bin/psql -c "ALTER USER postgres_exporter SET SEARCH_PATH TO postgres_exporter,pg_catalog;"
+    sudo -u postgres /usr/local/bin/psql -c "GRANT CONNECT ON DATABASE postgres TO postgres_exporter;"
+    sudo -u postgres /usr/local/bin/psql -c "GRANT pg_monitor to postgres_exporter;"
+else
+    echo "User postgres_exporter already exists. Not creating."
+fi