+++ postgres-single/CHANGELOG.md
@@ -8,3 +8,4 @@
 * Tabs and spaces fixes
 * Further adjustments to postgres_exporter user setup
 * Set a password for the postgres_exporter user
+* Add logic to check if postgresql users exist before adding